requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
Babel==2.12.1
backports.zoneinfo==0.2.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
filelock==3.11.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
g2pkk==0.1.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
idna==3.4
importlib-metadata==6.2.0
importlib-resources==5.12.0
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==1.2.0
kiwisolver==1.4.4
librosa==0.8.0
llvmlite==0.38.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mecab-python3==1.0.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.55.1
numpy==1.21.6
openai==0.27.4
packaging==23.0
pandas==2.0.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.7.0
protobuf==3.19.6
psutil==5.9.4
PyAudio==0.2.13
pycparser==2.21
pynndescent==0.5.8
pyparsing==3.0.9
pypinyin==0.48.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.0
sympy==1.11.1
tensorboardX==2.6
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
trainer==0.0.20
TTS==0.13.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
